Preparation for production
Created dockerfiles and docker compose, moved web and api ports to 15000 and 15001 respectively, implemented most of the functionality for ratings on the backend, few other small changes

+++ backend/api/views.py
@@ -2,9 +2,9 @@ from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.decorators import api_view, authentication_classes, permission_classes
 from rest_framework.permissions import IsAuthenticated
-from .models import Game, Tag, AuthToken, DiscordUser
-from .serializer import GameSerializer, TagSerializer, DiscordUserSerializer
-from .services import GameService
+from .models import Game, Tag, Rating
+from .serializer import GameSerializer, TagSerializer, DiscordUserSerializer, RatingSerializer
+from .services import get_user_from_auth_header
 from .auth import DiscordAuthBackend, DiscordTokenAuthentication
 
 # Endpoints
@@ -46,6 +46,15 @@ def deadgames(request):
     serializer = GameSerializer(games, many=True)
     return Response(serializer.data)
 
+# Get 6 random games
+@api_view(["GET"])
+@authentication_classes([DiscordTokenAuthentication])
+@permission_classes([IsAuthenticated])
+def randomgames(request):
+    games = Game.objects.all().order_by('?')[:6]
+    serializer = GameSerializer(games, many=True)
+    return Response(serializer.data)
+
 # Get game tags list
 @api_view(["GET"])
 @authentication_classes([DiscordTokenAuthentication])
@@ -84,7 +93,7 @@ def game(request, pk):
     
     # Update game
     elif request.method == "PATCH":
-        serializer = GameSerializer(game, data=request.data, partial=True)
+        serializer = GameSerializer(game, data=request.data, context=request.headers['Authorization'], partial=True)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
@@ -92,8 +101,37 @@ def game(request, pk):
     
     # Delete game
     elif request.method == "DELETE":
+        # Check permissions
+        game_added_by = game.added_by
+        user_object = get_user_from_auth_header(request.headers.get('Authorization'))
+        if user_object.id != game_added_by.id:
+            return Response({"error": "You do not have permission to delete this resource"}, status=status.HTTP_401_UNAUTHORIZED)
+
         game.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+# Ratings
+# Rate game
+@api_view(["POST"])
+@authentication_classes([DiscordTokenAuthentication])
+@permission_classes([IsAuthenticated])
+def add_rating(request):
+    serializer = RatingSerializer(data=request.data, context=request.headers['Authorization'])
+    if serializer.is_valid():
+        serializer.save()
+        return Response(serializer.data, status=status.HTTP_201_CREATED)
+    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+# Get current user's ratings
+@api_view(["GET"])
+@authentication_classes([DiscordTokenAuthentication])
+@permission_classes([IsAuthenticated])
+def user_ratings(request):
+    games = Rating.objects.filter(user=get_user_from_auth_header(request.headers.get('Authorization')))
+    serializer = RatingSerializer(games, many=True)
+    return Response(serializer.data)
+
 
 # Authentication & Users
 # Discord login
@@ -117,14 +155,12 @@ def discord_callback(request):
     })
 
 # Get currently logged in user
-@api_view(["GET", "PATCH", "DELETE"])
+@api_view(["GET"])
 @authentication_classes([DiscordTokenAuthentication])
 @permission_classes([IsAuthenticated])
 def current_user(request):
     # Get current user's token
-    token = request.headers.get('Authorization').split(' ')[1]
-    token_object = AuthToken.objects.get(token=token)
-    user_object = DiscordUser.objects.get(id=token_object.user_id)
+    user_object = get_user_from_auth_header(request.headers.get('Authorization'))
     
     # Return object
     serializer = DiscordUserSerializer(user_object)
